Extract app shortcut data clearing to remove duplicate code
